refactor: 重构组件路径，优化样式类名并添加新的控制组件
