+++ src/components/controls/StopButton.ts
