+++ src/components/controls/playPauseButton.ts
