+++ src/views/TexEditorView.ts
@@ -1,0 +1,1 @@
+export {};
