+++ styles.css
@@ -1,6 +1,6 @@
 /* styles.css */
 /* Add this class to your TabView's containerEl */
-.alphatab-obsidian-plugin .at-wrap {
+.itab .at-wrap {
   width: 100%;
   height: 100%;
   margin: 0 auto;
@@ -12,7 +12,7 @@
   box-sizing: border-box;
 }
 
-.alphatab-obsidian-plugin .at-content {
+.itab .at-content {
   position: relative;
   overflow: hidden;
   flex: 1 1 auto;
@@ -21,7 +21,7 @@
   flex-direction: column;
 }
 
-.alphatab-obsidian-plugin .at-viewport {
+.itab .at-viewport {
   overflow-y: auto;
   overflow-x: hidden; /* Usually you want vertical scroll for page layout */
   flex-grow: 1; /* Takes up available space in at-content */
@@ -32,12 +32,12 @@
 }
 
 /* Main AlphaTab rendering area */
-.alphatab-obsidian-plugin .at-main {
+.itab .at-main {
     width: 100%; /* Ensure it fills the viewport horizontally */
 }
 
 
-.alphatab-obsidian-plugin .at-controls {
+.itab .at-controls {
   flex: 0 0 auto; /* Do not grow or shrink, take needed height */
   padding: 8px;
   background-color: var(--background-secondary); /* Obsidian theme variable */
@@ -49,7 +49,7 @@
   flex-wrap: wrap; /* 允许控件换行显示 */
 }
 
-.alphatab-obsidian-plugin .at-controls button {
+.itab .at-controls button {
   /* Basic styling, Obsidian might style these by default */
   padding: 5px 10px;
   border-radius: 4px;
@@ -59,21 +59,21 @@
   cursor: pointer;
 }
 
-.alphatab-obsidian-plugin .at-controls button:hover {
+.itab .at-controls button:hover {
   background-color: var(--interactive-hover);
 }
 
-.alphatab-obsidian-plugin .at-controls button.active {
+.itab .at-controls button.active {
   background-color: var(--interactive-accent);
   color: var(--text-on-accent);
 }
 
-.alphatab-obsidian-plugin .at-controls button:disabled {
+.itab .at-controls button:disabled {
   opacity: 0.5;
   cursor: not-allowed;
 }
 
-.alphatab-obsidian-plugin .time-position {
+.itab .time-position {
   font-family: monospace;
   padding: 5px 10px;
   background-color: var(--background-primary);
@@ -81,15 +81,15 @@
   border: 1px solid var(--background-modifier-border);
 }
 
-.alphatab-obsidian-plugin .layout-control,
-.alphatab-obsidian-plugin .zoom-control,
-.alphatab-obsidian-plugin .speed-control {
+.itab .layout-control,
+.itab .zoom-control,
+.itab .speed-control {
   display: flex;
   align-items: center;
   gap: 4px;
 }
 
-.alphatab-obsidian-plugin .at-controls select {
+.itab .at-controls select {
   padding: 4px 8px;
   border-radius: 4px;
   border: 1px solid var(--background-modifier-border);
@@ -97,7 +97,7 @@
   color: var(--text-normal);
 }
 
-.alphatab-obsidian-plugin .at-overlay {
+.itab .at-overlay {
   position: absolute;
   top: 0;
   left: 0;
@@ -112,7 +112,7 @@
   box-sizing: border-box;
 }
 
-.alphatab-obsidian-plugin .at-overlay-content {
+.itab .at-overlay-content {
   background: var(--background-primary); /* Obsidian theme variable */
   color: var(--text-normal);
   box-shadow: 0px 5px 15px rgba(0, 0, 0, 0.2);
@@ -122,7 +122,7 @@
   box-sizing: border-box;
 }
 
-.alphatab-obsidian-plugin .at-overlay.error .at-overlay-content {
+.itab .at-overlay.error .at-overlay-content {
     background-color: var(--background-modifier-error);
     color: var(--text-on-error);
     border: 1px solid var(--color-red);
@@ -130,48 +130,48 @@
 
 /* AlphaTab Player Cursors/Highlights (from your Vue styles) */
 /* These need to target classes generated by AlphaTab within .at-main */
-.alphatab-obsidian-plugin .at-main .at-cursor-bar {
+.itab .at-main .at-cursor-bar {
   background-color: rgba(var(--color-yellow-rgb), 0.25) !important; /* Use Obsidian's yellow if available, or fallback */
 }
 
-.alphatab-obsidian-plugin .at-main .at-cursor-beat {
+.itab .at-main .at-cursor-beat {
   background-color: rgba(var(--color-blue-rgb), 0.75) !important; /* Use Obsidian's blue */
   /* width: 3px; */ /* AlphaTab usually controls this */
 }
 
-.alphatab-obsidian-plugin .at-main .at-highlight * { /* Target children of .at-highlight */
+.itab .at-main .at-highlight * { /* Target children of .at-highlight */
   fill: var(--color-accent) !important; /* Obsidian's accent color */
   stroke: var(--color-accent) !important;
 }
 
 /* Ensure Obsidian's default scrollbar styling is used or customize as needed */
-.alphatab-obsidian-plugin .at-viewport {
+.itab .at-viewport {
   scrollbar-width: thin;
   scrollbar-color: var(--scrollbar-thumb-bg) var(--scrollbar-bg);
 }
 
-.alphatab-obsidian-plugin .at-viewport::-webkit-scrollbar {
+.itab .at-viewport::-webkit-scrollbar {
   width: 8px;
   height: 8px;
 }
 
-.alphatab-obsidian-plugin .at-viewport::-webkit-scrollbar-thumb {
+.itab .at-viewport::-webkit-scrollbar-thumb {
   background-color: var(--scrollbar-thumb-bg);
   border-radius: 4px;
 }
 
-.alphatab-obsidian-plugin .at-viewport::-webkit-scrollbar-track {
+.itab .at-viewport::-webkit-scrollbar-track {
   background: var(--scrollbar-bg);
 }
 
 /* 添加新的布局容器样式 */
-.alphatab-obsidian-plugin .at-layout-container {
+.itab .at-layout-container {
   display: flex;
   width: 100%;
   height: 100%;
 }
 
-.alphatab-obsidian-plugin .at-main-content {
+.itab .at-main-content {
   flex: 1;
   display: flex;
   flex-direction: column;
@@ -179,7 +179,7 @@
 }
 
 /* 侧边栏样式 */
-.alphatab-obsidian-plugin .at-tracks-sidebar {
+.itab .at-tracks-sidebar {
   display: flex;
   flex-direction: column;
   width: 250px;
@@ -189,7 +189,7 @@
   overflow: hidden;
 }
 
-.alphatab-obsidian-plugin .at-sidebar-header {
+.itab .at-sidebar-header {
   display: flex;
   justify-content: space-between;
   align-items: center;
@@ -198,28 +198,28 @@
   background-color: var(--background-secondary-alt);
 }
 
-.alphatab-obsidian-plugin .at-sidebar-header h3 {
+.itab .at-sidebar-header h3 {
   margin: 0;
   font-size: 16px;
 }
 
-.alphatab-obsidian-plugin .at-sidebar-close {
+.itab .at-sidebar-close {
   cursor: pointer;
   opacity: 0.7;
 }
 
-.alphatab-obsidian-plugin .at-sidebar-close:hover {
+.itab .at-sidebar-close:hover {
   opacity: 1;
 }
 
-.alphatab-obsidian-plugin .at-sidebar-content {
+.itab .at-sidebar-content {
   flex: 1;
   overflow-y: auto;
   padding: 10px;
 }
 
 /* 移除footer样式，因为不再需要应用按钮
-.alphatab-obsidian-plugin .at-sidebar-footer {
+.itab .at-sidebar-footer {
   padding: 10px;
   border-top: 1px solid var(--background-modifier-border);
   display: flex;
@@ -227,19 +227,19 @@
 }
 */
 
-.alphatab-obsidian-plugin .at-sidebar-content .setting-item {
+.itab .at-sidebar-content .setting-item {
   padding: 10px 0;
   border-bottom: 1px solid var(--background-modifier-border-subtle);
 }
 
-.alphatab-obsidian-plugin .at-select-all-container {
+.itab .at-select-all-container {
   display: flex;
   justify-content: flex-start;
   margin-bottom: 10px;
   font-size: 0.9em;
 }
 
-.alphatab-obsidian-plugin .at-select-toggle-btn {
+.itab .at-select-toggle-btn {
   padding: 4px 12px;
   border-radius: 4px;
   background-color: var(--interactive-normal);
@@ -250,18 +250,18 @@
   transition: all 0.2s ease;
 }
 
-.alphatab-obsidian-plugin .at-select-toggle-btn:hover {
+.itab .at-select-toggle-btn:hover {
   background-color: var(--interactive-hover);
 }
 
-.alphatab-obsidian-plugin .at-select-toggle-btn.partial {
+.itab .at-select-toggle-btn.partial {
   background-color: var(--interactive-accent);
   color: var(--text-on-accent);
   opacity: 0.8;
 }
 
 /* 为"重置为默认"状态添加特殊样式 */
-.alphatab-obsidian-plugin .at-select-toggle-btn:has-text("重置为默认") {
+.itab .at-select-toggle-btn:has-text("重置为默认") {
   background-color: var(--color-orange);
   color: var(--text-on-accent);
 }
